refactor(ai): AiUsage 엔티티 및 Repository를 ai 도메인으로 이동
