+++ src/main/java/com/caro/bizkit/domain/ai/repository/AiUsageRepository.java
@@ -1,6 +1,6 @@
-package com.caro.bizkit.domain.user.repository;
+package com.caro.bizkit.domain.ai.repository;
 
-import com.caro.bizkit.domain.user.entity.AiUsage;
+import com.caro.bizkit.domain.ai.entity.AiUsage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.Optional;
 
 public interface AiUsageRepository extends JpaRepository<AiUsage, Integer> {
+
     void deleteByUserId(Integer userId);
 
     Optional<AiUsage> findByUserId(Integer userId);
