+++ src/main/java/com/caro/bizkit/domain/auth/service/AuthService.java
@@ -8,9 +8,9 @@ import com.caro.bizkit.domain.auth.entity.OAuth;
 import com.caro.bizkit.domain.auth.repository.AccountRepository;
 import com.caro.bizkit.domain.auth.repository.OAuthRepository;
 import com.caro.bizkit.domain.card.repository.CardRepository;
-import com.caro.bizkit.domain.user.entity.AiUsage;
+import com.caro.bizkit.domain.ai.entity.AiUsage;
 import com.caro.bizkit.domain.user.entity.User;
-import com.caro.bizkit.domain.user.repository.AiUsageRepository;
+import com.caro.bizkit.domain.ai.repository.AiUsageRepository;
 import com.caro.bizkit.domain.user.repository.UserRepository;
 import com.caro.bizkit.security.JwtTokenProvider;
 import com.caro.bizkit.security.RefreshTokenService;
